requirements.txt
torch
numpy
pyyaml
python-dotenv
requests
